Annotate the code that wraps around unsigned integers on purpose
The `offsets[i - 1]` idiom is the bulk of it. `PaddedPODArray` keeps zeroed
padding before the first element, which is why an array of offsets can be read
with `offsets[i - 1]` without a special case for the first element - its
subscript takes an `ssize_t`. But with an unsigned `i` the subtraction happens
before the call and wraps around. The index is now made signed first, which is
what a few places in the tree already did.

The rest of the places that wrap around on purpose - hash functions, checksums,
pseudo-random number generators, the delta codecs, DWARF and number parsing,
negative array indices, date/time arithmetic over the unsigned domain - say so
with `NO_SANITIZE_UNSIGNED_OVERFLOW`. The Keeper storage digest and the
`Replicated` database digest are rolling sums over their nodes, so they get
named `addToDigest`/`removeFromDigest` helpers instead of switching off the
check for whole storage functions.

A few places did not mean to wrap around at all:

* `ColumnIndex::insertIndexesRange` computed `start + length - 1` for an empty
  range, and the wrapped-around maximum index then widened the index type all
  the way.
* `MergeTreeIndexGranularityConstant::getRowsCountInRange` clamps `end` to the
  last constant-granularity mark, which can put it before `begin`; the product
  then wrapped into a nonsense row count.
* `ReadFromMergeTree` decremented the per-stream mark budget by the size of a
  part that can be bigger than the budget, so the counter wrapped instead of
  reaching zero and the stream went on to swallow every remaining part.
* `getCompressedDataSize` of the `DoubleDelta` codec computed the number of bits
  in 32 bits; above ~477 MB of 8-byte values it wrapped and returned a size
  several times too small. The same function was also called with a
  wrapped-around size for an input shorter than two items.
* `ZstdDeflatingWriteBuffer` measured its compressed output by subtracting
  buffer offsets taken before and after a loop that can flush the buffer in
  between and reset the offset.
* `SerializationNullable` read the element before the first one of an empty
  column to build a parse error message.
* Smaller ones in `ThreadPool::setMaxThreads`, `ProgressIndication`,
  `DecomposedFloat`, `arrayReduceInRanges`, `Volnitsky`, `reinterpretAs`,
  `ColumnTuple::popBack`, `readDecimalText` and a quantile binary search that
  only terminated because a later bound check caught the wrapped index.

The `i-- > 0` reverse-loop idiom wraps around on the last iteration by
construction; the ten places in our own code that used it decrement inside the
body instead.

One build change comes with it: the extra instrumentation pushes
`src/Interpreters/Aggregator.cpp` over the per-translation-unit memory limit of
sanitizer builds (`LLVM ERROR: out of memory` in the `Assignment Tracking
Analysis` pass). Measured on aarch64 with Clang 21, it peaks at 11.0 GB of
virtual memory with the check and 9.9 GB without, against a limit of 10 GB -
that is, the file has been sitting right at the limit for a while.

+++ src/AggregateFunctions/AggregateFunctionGroupConcat.cpp
@@ -200,8 +200,8 @@ void GroupConcatImpl<has_limit>::deserialize(AggregateDataPtr __restrict place, 
             if (cur_data.offsets[i] > cur_data.data_size)
                 throw Exception(ErrorCodes::BAD_ARGUMENTS, "Invalid offset {} in groupConcat state: exceeds data size {}", cur_data.offsets[i], cur_data.data_size);
 
-            if (i != 0 && cur_data.offsets[i] < cur_data.offsets[i - 1])
-                throw Exception(ErrorCodes::BAD_ARGUMENTS, "Invalid offsets in groupConcat state: end offset {} is less than start offset {}", cur_data.offsets[i], cur_data.offsets[i - 1]);
+            if (i != 0 && cur_data.offsets[i] < cur_data.offsets[static_cast<ssize_t>(i) - 1])
+                throw Exception(ErrorCodes::BAD_ARGUMENTS, "Invalid offsets in groupConcat state: end offset {} is less than start offset {}", cur_data.offsets[i], cur_data.offsets[static_cast<ssize_t>(i) - 1]);
         }
 
         if (cur_data.num_rows != 0 && cur_data.offsets.back() != cur_data.data_size)
