+++ src/Functions/formatRow.cpp
@@ -100,7 +100,7 @@ public:
                 /// `size_t` underflow in `ColumnString::sizeAt`. The check against `buffer.buffer().begin()`
                 /// additionally keeps the position within the current working buffer so `--buffer.position()`
                 /// never moves the cursor before it.
-                const size_t row_start = i == 0 ? 0 : offsets[i - 1];
+                const size_t row_start = i == 0 ? 0 : offsets[static_cast<ssize_t>(i) - 1];
                 if (buffer.count() > row_start && buffer.position() > buffer.buffer().begin()
                     && buffer.position()[-1] == '\n')
                     --buffer.position();
